Refactor events API code into separate files
