+++ docs/events.api.rst
@@ -6,6 +6,14 @@ events.api package
    :undoc-members:
    :show-inheritance:
 
+Subpackages
+-----------
+
+.. toctree::
+
+   events.api.serializers
+   events.api.viewsets
+
 Submodules
 ----------
 
@@ -17,26 +25,10 @@ events.api.permissions module
    :undoc-members:
    :show-inheritance:
 
-events.api.serializers module
------------------------------
-
-.. automodule:: events.api.serializers
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
 events.api.urls module
 ----------------------
 
 .. automodule:: events.api.urls
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-events.api.viewsets module
---------------------------
-
-.. automodule:: events.api.viewsets
    :members:
    :undoc-members:
    :show-inheritance:
