+++ docs/events.api.serializers.base.rst
@@ -1,0 +1,19 @@
+events.api.serializers.base package
+===================================
+
+.. automodule:: events.api.serializers.base
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Submodules
+----------
+
+events.api.serializers.base.calendarjs module
+---------------------------------------------
+
+.. automodule:: events.api.serializers.base.calendarjs
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
