+++ docs/events.api.serializers.event_registrations.rst
@@ -1,0 +1,19 @@
+events.api.serializers.event\_registrations package
+===================================================
+
+.. automodule:: events.api.serializers.event_registrations
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Submodules
+----------
+
+events.api.serializers.event\_registrations.list module
+-------------------------------------------------------
+
+.. automodule:: events.api.serializers.event_registrations.list
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
