+++ docs/events.api.serializers.events.rst
@@ -1,0 +1,35 @@
+events.api.serializers.events package
+=====================================
+
+.. automodule:: events.api.serializers.events
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Submodules
+----------
+
+events.api.serializers.events.calendarjs module
+-----------------------------------------------
+
+.. automodule:: events.api.serializers.events.calendarjs
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+events.api.serializers.events.list module
+-----------------------------------------
+
+.. automodule:: events.api.serializers.events.list
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+events.api.serializers.events.retrieve module
+---------------------------------------------
+
+.. automodule:: events.api.serializers.events.retrieve
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
