+++ docs/events.api.serializers.rst
@@ -1,0 +1,16 @@
+events.api.serializers package
+==============================
+
+.. automodule:: events.api.serializers
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Subpackages
+-----------
+
+.. toctree::
+
+   events.api.serializers.base
+   events.api.serializers.event_registrations
+   events.api.serializers.events
