+++ docs/events.api.viewsets.rst
@@ -1,0 +1,27 @@
+events.api.viewsets package
+===========================
+
+.. automodule:: events.api.viewsets
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Submodules
+----------
+
+events.api.viewsets.event\_registrations module
+-----------------------------------------------
+
+.. automodule:: events.api.viewsets.event_registrations
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+events.api.viewsets.events module
+---------------------------------
+
+.. automodule:: events.api.viewsets.events
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
