+++ website/events/api/serializers/__init__.py
@@ -1,0 +1,3 @@
+from .base import *
+from .events import *
+from .event_registrations import *
