+++ website/events/api/serializers/base/__init__.py
@@ -1,0 +1,1 @@
+from .calendarjs import *
