+++ website/events/api/serializers/event_registrations/__init__.py
@@ -1,0 +1,1 @@
+from .list import *
