+++ website/events/api/serializers/events/__init__.py
@@ -1,0 +1,3 @@
+from .calendarjs import *
+from .list import *
+from .retrieve import *
