+++ website/events/api/viewsets/__init__.py
@@ -1,0 +1,2 @@
+from .events import *
+from .event_registrations import *
